create client

+++ client.php
@@ -13,7 +13,7 @@
     <meta name="description" content="">
     <meta name="author" content="">
 
-    <title>AFG - Clients</title>
+    <title>AFC - Clients</title>
 
     <!-- Custom fonts for this template -->
     <link href="vendor/fontawesome-free/css/all.min.css" rel="stylesheet" type="text/css">
@@ -23,6 +23,7 @@
 
     <!-- Custom styles for this template -->
     <link href="css/sb-admin-2.min.css" rel="stylesheet">
+    
     <link rel="stylesheet" href="css/cient.css">
     <!-- Custom styles for this page -->
     <link href="vendor/datatables/dataTables.bootstrap4.min.css" rel="stylesheet">
@@ -36,118 +37,59 @@
 
         <!-- Sidebar -->
         <ul class="navbar-nav bg-gradient-primary sidebar sidebar-dark accordion" id="accordionSidebar">
-
-            <!-- Sidebar - Brand -->
-            <a class="sidebar-brand d-flex align-items-center" href="dashboard.php">
+                <!-- Sidebar - Brand -->
+                <!-- <a class="sidebar-brand d-flex align-items-center justify-content-center" href="dashboard.php">
+                    <div class="sidebar-brand-icon rotate-n-15">
+                        <i class="fas fa-laugh-wink"></i>
+                    </div>
+                    <div class="sidebar-brand-text mx-2">Africa Group</div>
+                </a> -->
+                <a class="sidebar-brand d-flex align-items-center" href="dashboard.php">
                     <!-- <div class="sidebar-brand-icon rotate-n-15">
                         <i class="fas fa-laugh-wink"></i>
                     </div> -->
                     <div class="sidebar-brand-text mx-2" style="font-size:1.25em;"><span style="font-size:1.5em;">A</span>FRIC'ACTION</div>
-            </a>
-
-            <!-- Divider -->
-            <hr class="sidebar-divider my-0">
-
-            <!-- Nav Item - Dashboard -->
-            <li class="nav-item">
-                <a class="nav-link" href="dashboard.php">
-                    <i class="fas fa-fw fa-tachometer-alt"></i>
-                    <span>Tableau de bord</span></a>
-            </li>
-
-            <!-- Divider -->
-            <hr class="sidebar-divider">
-
-            <!-- Heading -->
-            <div class="sidebar-heading">
-                Interface
-            </div>
-
-            <!-- Nav Item - Pages Collapse Menu -->
-            <li class="nav-item">
-                <a class="nav-link collapsed" href="#" data-toggle="collapse" data-target="#collapseTwo"
-                    aria-expanded="true" aria-controls="collapseTwo">
-                    <i class="fas fa-fw fa-cog"></i>
-                    <span>Components</span>
                 </a>
-                <div id="collapseTwo" class="collapse" aria-labelledby="headingTwo" data-parent="#accordionSidebar">
-                    <div class="bg-white py-2 collapse-inner rounded">
-                        <h6 class="collapse-header">Custom Components:</h6>
-                        <a class="collapse-item" href="buttons.html">Buttons</a>
-                        <a class="collapse-item" href="cards.html">Cards</a>
-                    </div>
+                <!-- Divider -->
+                <hr class="sidebar-divider my-0">
+                <!-- Nav Item - Dashboard -->
+                <li class="nav-item active">
+                    <a class="nav-link" href="dashboard.php">
+                        <i class="fas fa-fw fa-tachometer-alt"></i>
+                        <span>Tableau de bord</span>
+                    </a>
+                </li>
+
+                <!-- Divider -->
+                <hr class="sidebar-divider">
+                
+                <!-- Nav Item - Tables -->
+                <li class="nav-item">
+                    <a class="nav-link" href="client.php">
+                        <i class="fas fa-user"></i>
+                        <span>Clients</span></a>
+                </li>
+
+                <li class="nav-item">
+                    <a class="nav-link" href="abonnement.php">
+                        <i class="fas fa-clipboard-list"></i>
+                        <span>Abonnement</span></a>
+                </li>
+
+                <!-- Divider -->
+                <hr class="sidebar-divider d-none d-md-block">
+
+                <!-- Sidebar Toggler (Sidebar) -->
+                <div class="text-center d-none d-md-inline">
+                    <button class="rounded-circle border-0" id="sidebarToggle"></button>
                 </div>
-            </li>
 
-            <!-- Nav Item - Utilities Collapse Menu -->
-            <li class="nav-item">
-                <a class="nav-link collapsed" href="#" data-toggle="collapse" data-target="#collapseUtilities"
-                    aria-expanded="true" aria-controls="collapseUtilities">
-                    <i class="fas fa-fw fa-wrench"></i>
-                    <span>Utilities</span>
-                </a>
-                <div id="collapseUtilities" class="collapse" aria-labelledby="headingUtilities"
-                    data-parent="#accordionSidebar">
-                    <div class="bg-white py-2 collapse-inner rounded">
-                        <h6 class="collapse-header">Custom Utilities:</h6>
-                        <a class="collapse-item" href="utilities-color.html">Colors</a>
-                        <a class="collapse-item" href="utilities-border.html">Borders</a>
-                        <a class="collapse-item" href="utilities-animation.html">Animations</a>
-                        <a class="collapse-item" href="utilities-other.html">Other</a>
-                    </div>
-                </div>
-            </li>
-
-            <!-- Divider -->
-            <hr class="sidebar-divider">
-
-            <!-- Heading -->
-            <div class="sidebar-heading">
-                Addons
-            </div>
-
-            <!-- Nav Item - Pages Collapse Menu -->
-            <li class="nav-item">
-                <a class="nav-link collapsed" href="#" data-toggle="collapse" data-target="#collapsePages"
-                    aria-expanded="true" aria-controls="collapsePages">
-                    <i class="fas fa-fw fa-folder"></i>
-                    <span>Pages</span>
-                </a>
-                <div id="collapsePages" class="collapse" aria-labelledby="headingPages" data-parent="#accordionSidebar">
-                    <div class="bg-white py-2 collapse-inner rounded">
-                        <h6 class="collapse-header">Login Screens:</h6>
-                        <a class="collapse-item" href="login.html">Login</a>
-                        <a class="collapse-item" href="register.html">Register</a>
-                        <a class="collapse-item" href="forgot-password.html">Forgot Password</a>
-                        <div class="collapse-divider"></div>
-                        <h6 class="collapse-header">Other Pages:</h6>
-                        <a class="collapse-item" href="404.html">404 Page</a>
-                        <a class="collapse-item" href="blank.html">Blank Page</a>
-                    </div>
-                </div>
-            </li>
-
-            <!-- Nav Item - Charts -->
-            <li class="nav-item">
-                <a class="nav-link" href="charts.html">
-                    <i class="fas fa-fw fa-chart-area"></i>
-                    <span>Charts</span></a>
-            </li>
-
-            <!-- Nav Item - Tables -->
-            <li class="nav-item active">
-                <a class="nav-link" href="client.php">
-                    <i class="fas fa-fw fa-table"></i>
-                    <span>Clients</span></a>
-            </li>
-
-            <!-- Divider -->
-            <hr class="sidebar-divider d-none d-md-block">
-
-            <!-- Sidebar Toggler (Sidebar) -->
-            <div class="text-center d-none d-md-inline">
-                <button class="rounded-circle border-0" id="sidebarToggle"></button>
-            </div>
+                <!-- Sidebar Message -->
+                <!-- <div class="sidebar-card d-none d-lg-flex">
+                    <img class="sidebar-card-illustration mb-2" src="img/undraw_rocket.svg" alt="...">
+                    <p class="text-center mb-2"><strong>SB Admin Pro</strong> is packed with premium features, components, and more!</p>
+                    <a class="btn btn-success btn-sm" href="https://startbootstrap.com/theme/sb-admin-pro">Upgrade to Pro!</a>
+                </div> -->
 
         </ul>
         <!-- End of Sidebar -->
@@ -306,7 +248,7 @@
                         <!-- <div class="col-md-10"></div> -->
                         <div class="col-md-2">
                             <!-- <div class="mt-5 mb-3 clearfix"> -->
-                                <h2 class="pull-left" style="font-weight:bold;">Liste Clients</h2>
+                                <h2 class="pull-left" style="font-weight:bold;margin-left:20px;">Liste</h2>
                                 <!-- <a href="create.php" class="btn btn-sm btn-primary pull-right"><i class="fa fa-plus"></i> Add New Employee</a> -->
                             <!-- </div> -->
                             <!-- <a href="dashboard.php" class="btn btn-sm btn-primary" style="font-weight:bold;" id="NewClient">
@@ -546,12 +488,12 @@
                                                     <td class="actions">
                                                         <div class="btn-group">
                                                             <button class="btn btn-default dropdown-toggle" title='action' id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                                                            <i class="fa fa-ellipsis-v"></i>
+                                                            <i class="fas fa-ellipsis-v"></i>
                                                             </button>
                                                             <div class="dropdown-menu dropdown-menu-right" style="background-color:#F0F0F0;border:1px solid #F0F0F0">
                                                                 <a class="dropdown-item" href="seeModifyClient.php?id=<?php echo $row['idPersonne']?>" style="color:black;"><i class="fas fa-eye"></i>&nbsp;&nbsp;Voir</a>
-                                                                <div class="dropdown-divider" style="color:white;background-color:white;height:2px"></div>
-                                                                <a class="dropdown-item" href="deleteclient.php?id=<?php echo $row['idPersonne']?>" style="color:black"><i class="fas fa-trash-alt"></i>&nbsp;&nbsp;Supprimer</a>
+                                                                <div class="dropdown-divider" style="color:white;background-color:white;height:2.5px"></div>
+                                                                <a class="dropdown-item" href="deleteclient.php?id=<?php echo $row['idPersonne']?>" style="color:black" onclick="return confirm('Voulez vous vraiment supprimer ?')"><i class="fas fa-trash-alt"></i>&nbsp;&nbsp;Supprimer</a>
                                                             </div>
                                                         </div> 
                                                         <!-- <a href="#" class="btn btn-primary edit"><i class="fa fa-edit fa-xs"></i></a>
@@ -583,7 +525,7 @@
             <footer class="sticky-footer bg-white">
                 <div class="container my-auto">
                     <div class="copyright text-center my-auto">
-                        <span>Copyright &copy; Your Website 2020</span>
+                        <span>Copyright &copy; ariane.dalmeida 2022</span>
                     </div>
                 </div>
             </footer>
